+++ TabLayout/RefocusDecorator.js
@@ -42,7 +42,7 @@ function useScreenOrientation () {
 
 const RefocusDecorator = Wrapped => {
 	// eslint-disable-next-line no-shadow
-	function RefocusDecorator ({blockCollapseOnPortrait, collapsed, index, onCollapse, onTabAnimationEnd, orientation, spotlightId, ...rest}) {
+	function RefocusDecorator ({blockCollapseOnPortrait, blockExpandOnLandscape, collapsed, index, onCollapse, onExpand, onTabAnimationEnd, orientation, spotlightId, ...rest}) {
 		const {generateId} = useId({prefix: 'lime-tablayout-'});
 
 		const screenOrientation = useScreenOrientation();
@@ -72,20 +72,23 @@ const RefocusDecorator = Wrapped => {
 		}, [collapsed, orientation, spotlightId]);
 
 		useEffect(() => {
-			if (!blockCollapseOnPortrait && screenOrientationRef.current !== screenOrientation) {
-				if (screenOrientation === 'portrait') {
-					const currentFocusedElement = document.querySelector(':focus'),
-						tabsSpotlightId = getTabsSpotlightId(spotlightId, false),
-						tabsContainer = getContainerNode(tabsSpotlightId);
+			if (screenOrientationRef.current !== screenOrientation) {
+				const currentFocusedElement = document.querySelector(':focus'),
+					tabsSpotlightId = getTabsSpotlightId(spotlightId, false),
+					tabsContainer = getContainerNode(tabsSpotlightId);
 
-					if (tabsContainer && !tabsContainer.contains(currentFocusedElement)) {
+				if (tabsContainer && !tabsContainer.contains(currentFocusedElement)) {
+					if (!blockCollapseOnPortrait && screenOrientation === 'portrait') {
 						onCollapse();
+					}
+					if (!blockExpandOnLandscape && screenOrientation === 'landscape') {
+						onExpand();
 					}
 				}
 
 				screenOrientationRef.current = screenOrientation;
 			}
-		}, [blockCollapseOnPortrait, onCollapse, screenOrientation, spotlightId]);
+		}, [blockCollapseOnPortrait, blockExpandOnLandscape, onCollapse, onExpand, screenOrientation, spotlightId]);
 
 		const handleTabAnimationEnd = useCallback((ev) => {
 			if (onTabAnimationEnd) {
@@ -109,6 +112,7 @@ const RefocusDecorator = Wrapped => {
 				collapsed={collapsed}
 				index={index}
 				onCollapse={onCollapse}
+				onExpand={onExpand}
 				onTabAnimationEnd={handleTabAnimationEnd}
 				orientation={orientation}
 				spotlightId={spotlightId}