+++ samples/qa-a11y/src/views/RadioItem.js
@@ -1,4 +1,4 @@
-import RadioItem from '@enact/limestone/RadioItem';
+import RadioItem, {RadioItemGroup} from '@enact/limestone/RadioItem';
 
 import Section from '../components/Section';
 
@@ -14,6 +14,13 @@ const RadioItemView = () => (
 		<Section className={appCss.marginTop} title="Aria-labelled">
 			<RadioItem alt="Aria-labelled" aria-label="This is a Label 0.">Text 0</RadioItem>
 			<RadioItem alt="Aria-labelled and Disabled" aria-label="This is a Label 1." disabled>Text 1</RadioItem>
+		</Section>
+
+		<Section title="RadioItem Group">
+			<RadioItemGroup alt="Grouped">
+				<RadioItem>Text 0</RadioItem>
+				<RadioItem disabled>Text 1</RadioItem>
+			</RadioItemGroup>
 		</Section>
 	</>
 );